Move Retry with options to worbench-util
